backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
transformers==4.41.1
torch==2.3.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1